requirements.txt
azure-functions
msal
requests
python-dotenv
azure-search-documents
azure-core
azure-identity